Refactor button on Pricing page, Refactor pricing cards
