+++ src/components/GlobalStyles/Button.js
@@ -3,8 +3,8 @@ import { ButtonBack, ButtonFront } from './index'
 
 const Button = (props) => {
   return (
-    <ButtonBack>{props.children}
-      <ButtonFront onClick={props.onClick}>{props.children}</ButtonFront>
+    <ButtonBack alt={props.alt}>{props.children}
+      <ButtonFront alt={props.alt} onClick={props.onClick}>{props.children}</ButtonFront>
     </ButtonBack>
   )
 }
